d2 a-z

+++ Soal/Day 2/G.cpp
@@ -7,7 +7,7 @@ int main(){
 		int n;
 		scanf("%d",&n);
 		int score[n];
-		int avg=0;
+		double avg=0;
 		for(int i=0;i<n;i++){
 			scanf("%d",&score[i]);
 			avg+=score[i];
